requirements.txt
openai==0.27.8
Pillow==9.5.0
pytesseract==0.3.10
python-dotenv==0.19.2
google-cloud-texttospeech==2.14.1
moviepy==1.0.3
